Turned builtin options into proper objects.
