+++ meson.py
@@ -22,14 +22,11 @@ import build
 import platform
 import mlog, coredata
 
-from coredata import MesonException
-
-parser = argparse.ArgumentParser()
+from coredata import MesonException, build_types, layouts, warning_levels, libtypelist
 
 backendlist = ['ninja', 'vs2010', 'xcode']
-build_types = ['plain', 'debug', 'debugoptimized', 'release']
-layouts = ['mirror', 'flat']
-warning_levels = ['1', '2', '3']
+
+parser = argparse.ArgumentParser()
 
 default_warning = '2'
 
@@ -68,7 +65,7 @@ parser.add_argument('--werror', action='store_true', dest='werror', default=Fals
                     help='Treat warnings as errors')
 parser.add_argument('--layout', choices=layouts, dest='layout', default='mirror',\
                     help='Build directory layout.')
-parser.add_argument('--default-library', choices=['shared', 'static'], dest='default_library',
+parser.add_argument('--default-library', choices=libtypelist, dest='default_library',
                     default='static', help='Default library type.')
 parser.add_argument('--warnlevel', default=default_warning, dest='warning_level', choices=warning_levels,\
                     help='Level of compiler warnings to use (larger is more, default is %(default)s)')
